pyproject.toml
[project]
name = "office-admin"
version = "0.1.0"
description = "Asynchronous office automation sample using OfficeAdmin orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "certifi==2026.2.25",
    "cffi==2.0.0 ; platform_python_implementation != 'PyPy'",
    "charset-normalizer==3.4.7",
    "click==8.3.2",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "cryptography==46.0.7",
    "fastapi==0.136.0",
    "google-api-core==2.30.3",
    "google-api-python-client==2.194.0",
    "google-auth==2.49.2",
    "google-auth-httplib2==0.3.1",
    "google-auth-oauthlib==1.3.1",
    "googleapis-common-protos==1.74.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.31.2",
    "httpx==0.28.1",
    "idna==3.11",
    "iniconfig==2.3.0",
    "jinja2>=3.1.0,<4.0.0",
    "librt==0.9.0 ; platform_python_implementation != 'PyPy'",
    "mypy==1.20.1",
    "mypy-extensions==1.1.0",
    "oauthlib==3.3.1",
    "packaging==26.1",
    "pathspec==1.0.4",
    "pillow==12.2.0",
    "pluggy==1.6.0",
    "proto-plus==1.27.2",
    "protobuf==7.34.1",
    "pyasn1==0.6.3",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'",
    "pydantic==2.13.2",
    "pydantic-core==2.46.2",
    "pygments==2.20.0",
    "pyparsing==3.3.2",
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "reportlab==4.4.10",
    "requests==2.33.1",
    "requests-oauthlib==2.0.0",
    "ruff==0.15.11",
    "starlette==1.0.0",
    "types-reportlab==4.4.10.20260408",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
    "uvicorn==0.44.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0,<1.0.0",
    "pytest>=8.4.2,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
